+++ storage-node/CHANGELOG.md
@@ -1,3 +1,9 @@
+### 3.5.0
+- Multiple keys can now be passed as arguments to support multiple transactor accounts for node operation.
+- New `keyStore` argument can be used to pass a directory containing multiple key files.
+- `operatorId` renamed to `workerdId` for operator commands.
+- Server can now be run to operate specified buckets only.
+
 ### 3.4.0
 - Added option 'none' to 'logFileChangeFrequency' argument. The default is still 'daily'. 'none' prevents log rotaion on time basis and only rotates when max  size for logs files is reached.